refactor all the code related to groups
