+++ client/src/App.scss
@@ -44,13 +44,13 @@ a {
 button {
   background: #00a8e8;
   border-radius: 8px;
-  min-height: 34px;
+  min-height: 40px;
   min-width: 72px;
   font-size: 18px;
   font-weight: 500;
   color: #fff;
   border: #00a8e8;
-  line-height: 34px;
+  line-height: 40px;
   @include transition;
 
   &:hover {
