+++ client/src/Groups/Group.js
@@ -2,22 +2,9 @@ import React, { useState } from 'react';
 
 import Avatar from '../UI/Avatar';
 import Link from '../UI/Link';
-import { Done as AcceptIcon } from '../UI/icons';
-import { putRequest } from '../api';
+import Members from './Members';
+
 import { any } from '../array';
-
-const ACTIONS = { ACCEPT: 'ACCEPT', REJECT: 'REJECT', INVITE: 'INVITE' };
-
-const ActionButtons = ({ groupId, email }) => (
-  <div className="action-buttons">
-    <AcceptIcon
-      className="accept-icon"
-      onClick={() =>
-        putRequest(`groups/${groupId}`, { email, action: ACTIONS.ACCEPT })
-      }
-    />
-  </div>
-);
 
 export const getIdFromUri = () => {
   const { href } = location;
@@ -35,71 +22,6 @@ const handleChange = set => () => {
 
   set(value);
 };
-
-const Member = ({
-  currentUser: { _id: currentUserId },
-  invited,
-  groupId,
-  id: memberId,
-  index,
-  member: { name, email } = {}
-}) => {
-  const even = index % 2 === 0 ? 'even' : '';
-  const path = `/groups/${groupId}/members/${memberId}/`;
-  const maybeCurrentUser = currentUserId === memberId;
-
-  return (
-    <tr className={even}>
-      <td>{index}</td>
-      <td>
-        <Link path={path}>{name}</Link>
-      </td>
-      <td>{email}</td>
-      {maybeCurrentUser && invited && (
-        <td>
-          <ActionButtons email={email} groupId={groupId} />
-        </td>
-      )}
-    </tr>
-  );
-};
-
-const Members = ({
-  invited = false,
-  groupId,
-  membersIds,
-  users,
-  currentUser
-}) =>
-  membersIds && (
-    <div>
-      <h3>{invited ? 'Invited' : 'Members'}:</h3>
-      <table>
-        <thead>
-          <tr>
-            <th>#</th>
-            <th>Name</th>
-            <th>e-mail</th>
-          </tr>
-        </thead>
-        {membersIds.map((memberId, index) => {
-          const displayIndex = index + 1;
-
-          return (
-            <Member
-              currentUser={currentUser}
-              groupId={groupId}
-              id={memberId}
-              index={displayIndex}
-              invited={invited}
-              key={memberId}
-              member={users[memberId]}
-            />
-          );
-        })}
-      </table>
-    </div>
-  );
 
 const Group = ({ groups, users, currentUser }) => {
   const id = getIdFromUri();
